Common code changes to support `DocumentModel.getDisplayTitle()`
